+++ doc/CHANGES.rst
@@ -6,6 +6,7 @@
 ..    Copyright (C) 2018 Jean-Sebastien SUZANNE <jssuzanne@anybox.fr>
 ..    Copyright (C) 2019 Jean-Sebastien SUZANNE <js.suzanne@gmail.com>
 ..    Copyright (C) 2020 Jean-Sebastien SUZANNE <js.suzanne@gmail.com>
+..    Copyright (C) 2020 Pierre Verkest <pierreverkest84@gmail.com>
 ..
 .. This Source Code Form is subject to the terms of the Mozilla Public License,
 .. v. 2.0. If a copy of the MPL was not distributed with this file,You can
@@ -13,6 +14,12 @@
 
 CHANGELOG
 =========
+
+1.1.2 (unreleased)
+------------------
+
+* Added **restrict_query_by_user** decorator in order to apply query filters
+  related to a given user
 
 1.1.1 (2020-10-16)
 ------------------